Fixed for warnings: size_t/int issues, mainly
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@24813 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/imesh/snow.h
@@ -41,7 +41,7 @@ struct iSnowState : public virtual iBase
   /// Set the number of particles to use.
   virtual void SetParticleCount (int num) = 0;
   /// Get the number of particles used.
-  virtual int GetParticleCount () const = 0;
+  virtual size_t GetParticleCount () const = 0;
   /// Set the size of the drops.
   virtual void SetDropSize (float dropwidth, float dropheight) = 0;
   /// Get the size of the snow flakes.
