+++ plugins/mesh/spiral/persist/standard/spirldr.cpp
@@ -367,11 +367,11 @@ bool csSpiralSaver::WriteDown (iBase* obj, iDocumentNode* parent,
     synldr->WriteMixmode(mixmodeNode, mixmode, true);
 	  
     //Writedown Number tag
-    int number = spiralstate->GetParticleCount();
+    size_t number = spiralstate->GetParticleCount();
     csRef<iDocumentNode> numberNode = paramsNode->CreateNodeBefore(CS_NODE_ELEMENT, 0);
     numberNode->SetValue("number");
     csRef<iDocumentNode> numberValueNode = numberNode->CreateNodeBefore(CS_NODE_TEXT, 0);
-    numberValueNode->SetValueAsInt(number);
+    numberValueNode->SetValueAsInt((int)number);
   }
 
   paramsNode=0;
